refactor(settings): extract AlgorithmService
Moves the ~640-line "Algorithm helpers" block out of SettingsController
into src/settings/algorithmservice.{h,cpp}. Covers:

- Catalog queries: availableAlgorithms, generateAlgorithmPreview,
  generateAlgorithmDefaultPreview
- Filesystem CRUD: openAlgorithmsFolder, importAlgorithm, openAlgorithm,
  openLayoutFile, deleteAlgorithm, duplicateAlgorithm, exportAlgorithm,
  createNewAlgorithm (~200 lines on its own)
- Static utility algorithmIdFromLayoutId
- Private helpers: scriptedFilePath, watchForAlgorithmRegistration,
  cancelAlgorithmWatcher, the m_algorithmWatchers connection hash, and
  the anonymous-namespace userAlgorithmsDir / findUniqueAlgorithmPath
  file-location helpers

The service holds raw-pointer borrows of Settings / AlgorithmRegistry /
ScriptedAlgorithmLoader — no ownership transfer. Declared as
std::unique_ptr<AlgorithmService> AFTER m_scriptLoader in
SettingsController's member list, so reverse-order destruction runs the
service dtor (which disconnects pending watchers) BEFORE the
registry/loader unique_ptrs reset. SettingsController's explicit
watcher-disconnect loop in ~SettingsController is replaced by = default.

Signals (algorithmCreated, algorithmOperationFailed,
availableAlgorithmsChanged) emit from the service and are forwarded to
SettingsController's same-named signals via signal-to-signal connects in
the ctor. QML-facing signal names unchanged. The direct
ScriptedAlgorithmLoader::algorithmsChanged ↦ SettingsController::
availableAlgorithmsChanged wire moved into the service; the
layoutsChanged wire (separate concern) stays on SettingsController.

The 15 Q_INVOKABLE methods on SettingsController become 1-line
forwarders so QML entry points don't change.

settingscontroller.cpp: 2926 → 2337 lines (-589)
